Move PostgresFixture
