+++ client-dotnet/src/EvoMaster.Controller.Tests/Controllers/db/DbCleanerMSSqlServerTest.cs
@@ -7,7 +7,6 @@ using DotNet.Testcontainers.Containers.Configurations.Databases;
 using DotNet.Testcontainers.Containers.Modules.Databases;
 using EvoMaster.Controller.Api;
 using EvoMaster.Controller.Controllers.db;
-using Npgsql;
 using Xunit;
 
 namespace EvoMaster.Controller.Tests.Controllers.db {
@@ -46,42 +45,6 @@ namespace EvoMaster.Controller.Tests.Controllers.db {
 
         protected override void CleanDb(List<string> tablesToSkip) {
             DbCleaner.ClearDatabase(_connection, tablesToSkip, GetDbType(), "");
-        }
-    }
-    public class PostgresFixture : IAsyncLifetime
-    {
-
-        private DbConnection _connection;
-        private PostgreSqlTestcontainer _postgres;
-
-        public async Task InitializeAsync()
-        {
-            //for the moment, use this testcontainer for dotnet https://github.com/HofmeisterAn/dotnet-testcontainers
-            ITestcontainersBuilder<PostgreSqlTestcontainer> postgresBuilder =
-                new TestcontainersBuilder<PostgreSqlTestcontainer>()
-                    .WithDatabase(new PostgreSqlTestcontainerConfiguration("postgres:11.5")
-                    {
-                        Database = "db",
-                        Username = "postgres",
-                        Password = "postgres",
-                    })
-                    .WithExposedPort(5432);
-            
-            _postgres = postgresBuilder.Build();
-            await _postgres.StartAsync();
-            _connection = new NpgsqlConnection(_postgres.ConnectionString);
-            await _connection.OpenAsync();
-        }
-
-        public async Task DisposeAsync()
-        {
-            await _connection.CloseAsync();
-            await _postgres.StopAsync();
-        }
-
-        public DbConnection GetConnection()
-        {
-            return _connection;
         }
     }
 }
